Use int types for B&B matrices and bounds
Convert Branch-and-Bound matrix and bound types from long long/LLONG_MAX to int/numeric_limits<int>::max(): update reduceMatrix to accept/return int matrices, change Node.lowerBound to int, and adjust branchAndBoundBFS/DFS/BEST signatures and internal matrix/storage types accordingly. Also update main to use an int initial UB, tweak badania.bat loop range, replace config defaults (algorithm, input/output, sizes), and add many results CSV files. These changes are primarily type/configuration adjustments (no algorithmic logic changes other than type conversions).

+++ algorithms.h
@@ -17,7 +17,7 @@ int bruteForceTSP(const std::vector<std::vector<int>>& graph);
 // Optimized Node structure for Branch and Bound with reference semantics
 struct Node {
     std::vector<int> path;
-    long long lowerBound;
+    int lowerBound;
     int level;
 
     Node(int n) : path(), lowerBound(0), level(0) {}
@@ -35,11 +35,11 @@ struct Node {
 };
 
 // Branch and Bound implementation
-int branchAndBoundBFS(const std::vector<std::vector<int>>& graph, long long initialUB);
-int branchAndBoundDFS(const std::vector<std::vector<int>>& graph, long long initialUB);
-int branchAndBoundBEST(const std::vector<std::vector<int>>& graph, long long initialUB);
+int branchAndBoundBFS(const std::vector<std::vector<int>>& graph, int initialUB);
+int branchAndBoundDFS(const std::vector<std::vector<int>>& graph, int initialUB);
+int branchAndBoundBEST(const std::vector<std::vector<int>>& graph, int initialUB);
 
 // Helper function for matrix reduction
-long long reduceMatrix(std::vector<std::vector<long long>>& matrix);
+int reduceMatrix(std::vector<std::vector<int>>& matrix);
 
 #endif
